+++ hvad/descriptors.py
@@ -1,7 +1,7 @@
 import django
 from django.db.models.fields import FieldDoesNotExist
 from django.utils.translation import get_language
-from hvad.utils import get_translation
+from hvad.utils import get_translation, set_cached_translation, get_cached_translation
 if django.VERSION >= (1, 7):
     from django.apps import registry
 
@@ -16,29 +16,26 @@ class BaseDescriptor(object):
                                         {})
     
     def translation(self, instance):
-        cached = getattr(instance, self.opts.translations_cache, None)
-        if cached is None:
+        translation = get_cached_translation(instance)
+        if translation is None:
             try:
-                cached = get_translation(instance)
+                translation = get_translation(instance)
             except self.opts.translations_model.DoesNotExist:
                 raise self._NoTranslationError('Accessing a translated field requires that '
                                                'the instance has a translation loaded, or a '
                                                'valid translation in current language (%s) '
                                                'loadable from the database' % get_language())
-            setattr(instance, self.opts.translations_cache, cached)
-        return cached
+            set_cached_translation(instance, translation)
+        return translation
 
 
 class TranslatedAttribute(BaseDescriptor):
-    """
-    Basic translated attribute descriptor.
-    
-    Proxies attributes from the shared instance to the translated instance.
-    """
+    """ Proxies attributes from the shared instance to the translated instance. """
+
     def __init__(self, opts, name):
         self.name = name
         super(TranslatedAttribute, self).__init__(opts)
-        
+
     def __get__(self, instance, instance_type=None):
         if not instance:
             if django.VERSION >= (1, 7) and not registry.apps.ready:
@@ -68,8 +65,8 @@ class LanguageCodeAttribute(TranslatedAttribute):
         super(LanguageCodeAttribute, self).__init__(opts, 'language_code')
     
     def __set__(self, instance, value):
-        raise AttributeError("The 'language_code' attribute cannot be " +\
-                    "changed directly! Use the translate() method instead.")
+        raise AttributeError("The 'language_code' attribute cannot be "
+                             "changed directly! Use the translate() method instead.")
     
     def __delete__(self, instance):
         raise AttributeError("The 'language_code' attribute cannot be deleted!")